Ability to tear down Image Builder instance (#35)
## Why is this change necessary?
Want to be able to delete the EC2 instance used to build the AMI while
still preserving the information about how it was originally configured


 ## How does this change address the issue?
Adds a kwarg to mark delete it. Uses some conditional stuff to still
allow the AMI creation resource to exist.


 ## What side effects does this change have?
None


 ## How is this change tested?
Local development in downstream repo